Conda: Enable use of mingw-w64 g++ and mkl when not on PATH
Many Windows users avoid the command-line and are unfamiliar with env. vars.
They can elect to use Anaconda/Miniconda Python as their system Python and
in that case would not be able to take advantage of g++ or mkl.

This commit works around these issues by adding the appropriate values to
os.environ['PATH'] if they are not already on PATH.